Update heuristic_fractions.py
The updated code with the max function to select last N labels produced  output where the whole list of screened labels was returned rather than the last N. I moved it to the except statement in case the current method fails downstream. The fixed version was tested on a few datasets, but _disclaimer_ is not much different to the original code.